% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareX.R
\name{check_Xy}
\alias{check_Xy}
\title{Joint validation/preprocessing of `(X, y)`.}
\usage{
check_Xy(X, y, y_kind, standardize_X = TRUE, X_mean = NULL, X_std = NULL)
}
\description{
For survival data, rows of `X` and `y` are sorted by ascending time.
}
\keyword{internal}
