% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.pic.pdb_asymptotic}
\alias{print.pic.pdb_asymptotic}
\title{Print PDB asymptotic diagnostic.}
\usage{
\method{print}{pic.pdb_asymptotic}(x, ...)
}
\arguments{
\item{x}{A `pic.pdb_asymptotic` object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns `x`.
}
\description{
Print PDB asymptotic diagnostic.
}
