% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.pic.phase_transition}
\alias{print.pic.phase_transition}
\title{Print phase-transition analysis.}
\usage{
\method{print}{pic.phase_transition}(x, ...)
}
\arguments{
\item{x}{A \code{pic.phase_transition} object.}

\item{...}{Unused; present for S3 method consistency.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print phase-transition analysis.
}
