% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-picr.R
\name{print.summary.pic}
\alias{print.summary.pic}
\title{Print a \code{summary.pic} object.}
\usage{
\method{print}{summary.pic}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{summary.pic} object.}

\item{digits}{Number of significant digits for the printed values.}

\item{...}{Unused; present for S3 method consistency.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{summary.pic} object.
}
