% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picreg-package.R
\docType{package}
\name{picreg-package}
\alias{picreg}
\alias{picreg-package}
\title{picreg: Variable Selection using the Pivotal Information Criterion}
\description{
Sparse regression and classification with automatic regularisation
parameter selection via the Pivotal Detection Boundary (PDB) method.
Implements FISTA optimisation for L1, SCAD and MCP penalties across
Gaussian, Binomial, Poisson, Exponential, Gumbel and Cox families.
}
\seealso{
Useful links:
\itemize{
  \item <TODO: project URL
  \item e.g. \url{https://github.com/MaximeVanCutsem/picreg>}
  \item Report bugs at \url{<TODO: issue tracker URL, e.g. https://github.com/MaximeVanCutsem/picreg/issues>}
}

}
\author{
\strong{Maintainer}: Maxime Van Cutsem \email{maxime.vancutsem@unige.ch}

Authors:
\itemize{
  \item Maxime Van Cutsem \email{maxime.vancutsem@unige.ch}
}

}
\keyword{internal}
