% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-picr.R
\name{predict.pic}
\alias{predict.pic}
\title{Linear predictor / response prediction for a pic fit.}
\usage{
\method{predict}{pic}(object, newx, type = c("response", "link", "class"), ...)
}
\arguments{
\item{object}{A fitted `pic` object.}

\item{newx}{Matrix of new values at which predictions are to be made.}

\item{type}{One of `"link"` (linear predictor),
`"response"` (default; family `g`-link applied),
`"coefficients"` (return fitted coefficients incl. intercept)}

\item{...}{Ignored.}
}
\value{
A numeric vector.
}
\description{
Linear predictor / response prediction for a pic fit.
}
