% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.pic.phase_transition}
\alias{print.pic.phase_transition}
\title{Print phase-transition analysis.}
\usage{
\method{print}{pic.phase_transition}(x, ...)
}
\arguments{
\item{x}{A `pic.phase_transition` object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns `x`.
}
\description{
Print phase-transition analysis.
}
